refactor(form): move test files location
